backend/requirements.txt
asgiref==3.5.2
attrs==22.1.0
black==22.6.0
Brotli==1.0.9
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
cssselect2==0.6.0
defusedxml==0.7.1
Django==4.0.6
django-cleanup==6.0.0
django-colorfield==0.7.2
django-cors-headers==3.13.0
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.4.0
drf-yasg==1.21.3
flake8==5.0.4
fonttools==4.36.0
gunicorn==20.1.0
html5lib==1.1
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.9.1
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
oauthlib==3.2.0
packaging==21.3
pathspec==0.9.0
Pillow==9.2.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pycparser==2.21
pydyf==0.2.0
pyflakes==2.5.0
PyJWT==2.4.0
pyparsing==3.0.9
pyphen==0.12.0
pyrsistent==0.18.1
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
swagger-spec-validator==2.7.4
tinycss2==1.1.1
tomli==2.0.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.11
weasyprint==56.1
webencodings==0.5.1
zopfli==0.2.1
